+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Move `enableFileManagerSwizzling` from experimental options to top-level options (#6592).
   This option is still disabled by default and will be enabled in a future major release.
 - Move `enableDataSwizzling` from experimental options to top-level options (#6592). This option remains enabled by default.
+- Add `sentry.replay_id` attribute to logs ([#6515](https://github.com/getsentry/sentry-cocoa/pull/6515))
 
 ### Fixes